@@ -2,7 +2,7 @@ export default {
     en: {
         Index: {
             title: 'Team of developers',
-            desc: 'SuperB development with\u00a01C\u2011Bitrix, Laravel, Symfony, ReactJS',
+            desc: 'SuperB development with\u00a01C\u2011Bitrix, Symfony, ReactJS',
             anyQuestions: 'If you have any questions',
             agenciesLabel: 'projects from digital agencies',
             clientsLabel: 'projects from clients',
@@ -26,7 +26,7 @@ export default {
     ru: {
         Index: {
             title: 'Команда разработчиков',
-            desc: 'Профессиональная разработка сайтов на\u00a01C\u2011Битрикс, Laravel, Symfony, ReactJS',
+            desc: 'Профессиональная разработка сайтов на\u00a01C\u2011Битрикс, Symfony, ReactJS',
             anyQuestions: 'По любым вопросам',
             agenciesLabel: 'проектов по заказу агентств',
             clientsLabel: 'проектов по заказу клиентов',
