@@ -7,6 +7,7 @@ import { createStore, compose, applyMiddleware } from 'redux';
 import { Provider } from 'react-redux';
 import { Router, browserHistory } from 'react-router';
 import { syncHistoryWithStore } from 'react-router-redux';
+import { createLogger } from 'redux-logger';
 import createRoutes from './routes';
 import reducers from './state';
 import DevTools from './components/DevTools';
@@ -15,10 +16,9 @@ const initialState = JSON.parse(document.getElementById('app').getAttribute('dat
 
 let store;
 if (__DEVMODE__) {
-    const createLogger = require('redux-logger');
     const logger = createLogger({ predicate:
         (getState, action) => action.type !== 'EFFECT_TRIGGERED' &&
-        action.type !== 'EFFECT_RESOLVED' });
+                              action.type !== 'EFFECT_RESOLVED' });
     const enhance = compose(applyMiddleware(logger), DevTools.instrument());
 
     store = createStore(reducers, initialState, enhance);
